Authentication Implemented for real, this time. Lol

+++ src/api/albums/handler.js
@@ -1,5 +1,6 @@
 const _autoBind = require('auto-bind');
 const autoBind = _autoBind.default ?? _autoBind;
+const AuthenticationError = require('../../exceptions/AuthenticationError');
 
 class AlbumsHandler {
   constructor(service, validator) {
@@ -10,6 +11,11 @@ class AlbumsHandler {
 
   async postAlbumHandler(request, h) {
     this._validator.validateAlbumPayload(request.payload);
+    
+    if (!request.auth.isAuthenticated) {
+      throw new AuthenticationError('Missing authentication');
+    }
+    
     const { name, year, genre, performer } = request.payload;
 
     const albumId = await this._service.addAlbum({
@@ -37,27 +43,40 @@ class AlbumsHandler {
     };
   }
 
-  async putAlbumByIdHandler(request) {
+  async putAlbumByIdHandler(request, h) {
     this._validator.validateAlbumPayload(request.payload);
+    
+    if (!request.auth.isAuthenticated) {
+      throw new AuthenticationError('Missing authentication');
+    }
+    
     const { id } = request.params;
     const { name, year, genre, performer } = request.payload;
 
     await this._service.editAlbumById(id, { name, year, genre, performer });
 
-    return {
+    const response = h.response({
       status: 'success',
       message: 'Album berhasil diperbarui',
-    };
+    });
+    response.code(200);
+    return response;
   }
 
-  async deleteAlbumByIdHandler(request) {
+  async deleteAlbumByIdHandler(request, h) {
+    if (!request.auth.isAuthenticated) {
+      throw new AuthenticationError('Missing authentication');
+    }
+    
     const { id } = request.params;
     await this._service.deleteAlbumById(id);
 
-    return {
+    const response = h.response({
       status: 'success',
       message: 'Album berhasil dihapus',
-    };
+    });
+    response.code(200);
+    return response;
   }
 }
